+++ base/src/main/java/proguard/classfile/JavaVersionConstants.java
@@ -50,4 +50,5 @@ public class JavaVersionConstants
     public static final String CLASS_VERSION_16        = "16";
     public static final String CLASS_VERSION_17        = "17";
     public static final String CLASS_VERSION_18        = "18";
+    public static final String CLASS_VERSION_19        = "19";
 }